# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Service the secret santa website.",
        "title": "Service Secret Santa",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/secret-santa",
    "paths": {
        "/group": {
            "get": {
                "description": "Retrieve a list of all groups",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "group"
                ],
                "summary": "Get all groups",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Group"
                            }
                        }
                    },
                    "500": {
                        "description": "{\"error\": \"Internal Server Error.\"}"
                    }
                }
            },
            "post": {
                "description": "Create a new group with a name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "group"
                ],
                "summary": "Create a new group",
                "parameters": [
                    {
                        "description": "Group object",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Group"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Group"
                        }
                    },
                    "400": {
                        "description": "{\"error\": \"Bad Request.\"}"
                    },
                    "500": {
                        "description": "{\"error\": \"Internal Server Error.\"}"
                    }
                }
            }
        },
        "/group/{id}": {
            "get": {
                "description": "Retrieve details of a specific group by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "group"
                ],
                "summary": "Get a group by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Group ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Group"
                        }
                    },
                    "404": {
                        "description": "{\"error\": \"Not Found.\"}"
                    },
                    "500": {
                        "description": "{\"error\": \"Internal Server Error.\"}"
                    }
                }
            },
            "put": {
                "description": "Update details of a specific group",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "group"
                ],
                "summary": "Update a group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Group ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated group object",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Group"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Group"
                        }
                    },
                    "400": {
                        "description": "{\"error\": \"Bad Request.\"}"
                    },
                    "404": {
                        "description": "{\"error\": \"Not Found.\"}"
                    },
                    "500": {
                        "description": "{\"error\": \"Internal Server Error.\"}"
                    }
                }
            },
            "delete": {
                "description": "Delete a specific group by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "group"
                ],
                "summary": "Delete a group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Group ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "{}"
                    },
                    "404": {
                        "description": "{\"error\": \"Not Found.\"}"
                    },
                    "500": {
                        "description": "{\"error\": \"Internal Server Error.\"}"
                    }
                }
            }
        },
        "/group/{id}/add-participant": {
            "post": {
                "description": "Add a new participant to an existing group",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "group"
                ],
                "summary": "Add a participant to a group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Group ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Participant to add",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AddParticipantRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Group"
                        }
                    },
                    "400": {
                        "description": "{\"error\": \"Bad Request.\"}"
                    },
                    "404": {
                        "description": "{\"error\": \"Not Found.\"}"
                    },
                    "500": {
                        "description": "{\"error\": \"Internal Server Error.\"}"
                    }
                }
            }
        },
        "/group/{id}/match-participants": {
            "post": {
                "description": "Generate secret matches for participants in a group",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "group"
                ],
                "summary": "Match participants in a group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Group ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Group"
                        }
                    },
                    "400": {
                        "description": "{\"error\": \"Bad Request.\"}"
                    },
                    "404": {
                        "description": "{\"error\": \"Not Found.\"}"
                    },
                    "500": {
                        "description": "{\"error\": \"Internal Server Error.\"}"
                    }
                }
            }
        },
        "/group/{id}/my-match": {
            "get": {
                "description": "Retrieve the participant you are matched to gift in a group",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "group"
                ],
                "summary": "Get the match for a participant",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Group ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Participant username",
                        "name": "username",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "string"
                    },
                    "400": {
                        "description": "{\"error\": \"Bad Request.\"}"
                    },
                    "404": {
                        "description": "{\"error\": \"Not Found.\"}"
                    },
                    "500": {
                        "description": "{\"error\": \"Internal Server Error.\"}"
                    }
                }
            }
        }
    },
    "definitions": {
        "models.AddParticipantRequest": {
            "type": "object",
            "required": [
                "participant"
            ],
            "properties": {
                "participant": {
                    "type": "string"
                }
            }
        },
        "models.Group": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "Equipe pe no chao"
                },
                "participants": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "joao",
                        "mari"
                    ]
                }
            }
        }
    },
    "externalDocs": {
        "description": "ReadMe"
    }
}